Fix sstv

+++ addons/scripts/sstv.sh
@@ -27,13 +27,22 @@ UNIXTD="${3:-$(date -u +%s)}" # date -d "2024-04-25T14:07:37" -u +%s
 SATNAME=$(echo "$TLE" | jq .tle0 | sed -e 's/ /_/g' | sed -e 's/[^A-Za-z0-9._-]//g')
 NORAD=$(echo "$TLE" | jq .tle2 | awk '{print $2}')
 
+if [ -z "$CMD" ] || [ -z "$ID" ] || [ -z "$FREQ" ] || [ -z "$TLE" ]; then
+  echo "$PRG Error: missing variables"
+  exit 1
+fi
+
 if [ "${CMD^^}" = "START" ]; then
   if [[ "${MODE,,}" =~ "sstv" ]]; then
-    OPT="-d $SATNOGS_APP_PATH/satnogs_$ID_$DATE_OBS.ogg -o $SATNOGS_APP_PATH/sstv_$ID.png"
-    if [ -n "$OPT" ]; then
-      echo "$PRG $OPT"
-      $BIN $OPT
+    DATE_OBS=$(date +"%Y-%m-%dT%H-%M-%S")
+    OGG="${SATNOGS_APP_PATH}/satnogs_${ID}_${DATE_OBS}.ogg"
+    OPT="-d \"$OGG\" -o \"$SATNOGS_APP_PATH/sstv_$ID.png\""
+    echo "$PRG $OPT"
+    if [ -z "$BIN" ]; then
+      echo "$PRG Error: sstv command not found"
+      exit 1
     fi
+
     if [ -f "$SATNOGS_APP_PATH/sstv_$ID.png" ]; then
       echo "$PRG Processing data $OUT to network"
       # find images, rename/move to ${SATNOGS_OUTPUT_PATH}/data_<obsid>_YYYY-MM-DDTHH-MM-SS.png
@@ -41,15 +50,14 @@ if [ "${CMD^^}" = "START" ]; then
       month=$(date +"%m")
       day=$(date "+%d")
       hour=$(date "+%H")
-      DATE_OBS=$(date +"%Y-%m-%dT%H-%M-%S")
       basename_dest="${SATNOGS_OUTPUT_PATH}/data_${ID}_${DATE_OBS}_sstv.png"
       if cp "$SATNOGS_APP_PATH/sstv_$ID.png" "$basename_dest"; then
         echo "$PRG The image $basename_dest was transferred to the Satnogs network"
-        if [ "${SATNOGS_REMOVE_OGG_FILES^^}" = "TRUE" ]; then
+        if [ -f "$SATNOGS_APP_PATH/sstv_$ID.png" ] && [ "${SATNOGS_REMOVE_OGG_FILES^^}" = "TRUE" ]; then
           rm "$SATNOGS_APP_PATH/sstv_$ID.png"
         fi
       else
-        echo "$PRG Error transferring the image $file"
+        echo "$PRG Error transferring the image ${SATNOGS_APP_PATH}/sstv_${ID}.png"
       fi
     else
       echo "$PRG No SSTV image gererated !"
